Adapt to new payson web page
Payson was broken in bankdroid because of changes to payson's web page. Adapt.

Support for listing transactions is not yet implemented. The transactions
are no longer available in json format, but must instead be parsed from HTML.
