+++ bankdroid-legacy/src/main/res/values-se/strings.xml
@@ -7,6 +7,7 @@
     <string name="email">E-post</string>
     <string name="points">poäng</string>
     <string name="card_id">Kort ID</string>
+    <string name="balance">Saldo</string>
 
     <string name="nordnetdirekt_extras_title">Nyckel</string>
