+++ bankdroid-legacy/src/main/res/values/strings.xml
@@ -11,6 +11,7 @@
     <string name="email">E-mail</string>
     <string name="points">points</string>
     <string name="card_id">Card ID</string>
+    <string name="balance">Balance</string>
 
     <string name="nordnetdirekt_extras_title">Key</string>
